Remove git blame from CommandRunner's reponsibility
